Use return value instead of exception to indicate call success
I.e. we now return an object convertible to bool from the call operator.
If we fail to acquire a lock on the LockablePtr that object will convert
to false, instead of throwing std::bad_function_call. When the
designated return type is not void this object can additionally be
dereferenced to retrieve the returned value.